separate households test code from clients test file
